+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Apps
 - [radar] Enforce minimum constraint on size of tab text
 - [radar] Add `+` and `-` for zooming the map during operation
+- [radar] Add `Up`, `Down`, `Left`, and `Right` for moving map (lat/long). `Enter` for resetting Map.
+- [radar] Display current Lat/Long in tui
 
 ## [v0.2.0] 2021-10-17
 ### adsb_deku